add directories to components directory and moves files
