+++ src/components/features/AboutChessWar.vue
